# apps/api/openapi.yaml
openapi: 3.0.3
info:
  title: CouplePlan API
  version: 0.1.0
  description: MVP API contract for P1-1 (auth and couple pairing)
servers:
  - url: https://api-staging.example.com
    description: Staging
  - url: https://api.example.com
    description: Production
tags:
  - name: auth
  - name: users
  - name: couples
paths:
  /auth/otp/request:
    post:
      tags: [auth]
      summary: Request OTP code by email
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpRequest"
      responses:
        "202":
          description: OTP request accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OtpRequestAccepted"
  /auth/otp/verify:
    post:
      tags: [auth]
      summary: Verify OTP and issue session token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OtpVerifyRequest"
      responses:
        "200":
          description: OTP verified
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
  /auth/refresh:
    post:
      tags: [auth]
      summary: Refresh access token from HttpOnly refresh cookie
      responses:
        "200":
          description: Session refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthSession"
  /auth/logout:
    post:
      tags: [auth]
      summary: Revoke refresh token and clear auth cookie
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok:
                    type: boolean
  /users/me:
    get:
      tags: [users]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      tags: [users]
      summary: Update current user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserProfileUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      tags: [users]
      summary: Withdraw current account and delete couple-scoped data
      description: Deletes the authenticated user. If the user belongs to a couple, all users and shared data in that couple are deleted.
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Account data deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountDeletion"
  /couples:
    post:
      tags: [couples]
      summary: Create couple in pending state
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Couple created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Couple"
  /couples/me:
    get:
      tags: [couples]
      summary: Get current user's couple relation
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Couple found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Couple"
        "404":
          description: Couple not found
  /couples/invites:
    post:
      tags: [couples]
      summary: Issue invite code for current couple
      security:
        - bearerAuth: []
      responses:
        "201":
          description: Invite code issued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Invite"
  /couples/invites/{code}/accept:
    post:
      tags: [couples]
      summary: Accept invite code and activate couple
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Invite accepted and couple activated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Couple"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    OtpRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
    OtpRequestAccepted:
      type: object
      required: [requestId, expiresInSec]
      properties:
        requestId:
          type: string
        expiresInSec:
          type: integer
          example: 600
        debugCode:
          type: string
          description: 開発/メール未接続/ALLOW_DEBUG 時のみ。本番＋本番配信利用時は返さない
    OtpVerifyRequest:
      type: object
      required: [email, code]
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          minLength: 4
          maxLength: 8
    AuthSession:
      type: object
      required: [accessToken, user]
      properties:
        accessToken:
          type: string
        refreshExpiresInSec:
          type: integer
          description: Returned on OTP verify when refresh cookie is issued.
        user:
          $ref: "#/components/schemas/User"
    User:
      type: object
      required: [id, email, displayName, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
          format: email
        displayName:
          type: string
        createdAt:
          type: string
          format: date-time
    UserProfileUpdate:
      type: object
      properties:
        displayName:
          type: string
    AccountDeletion:
      type: object
      required: [deleted, deletedUserIds]
      properties:
        deleted:
          type: boolean
          enum: [true]
        deletedUserIds:
          type: array
          items:
            type: string
        deletedCoupleId:
          type: string
    Couple:
      type: object
      required: [id, status, members]
      properties:
        id:
          type: string
        status:
          type: string
          enum: [pending, active, unpaired]
        members:
          type: array
          items:
            $ref: "#/components/schemas/CoupleMember"
    CoupleMember:
      type: object
      required: [userId, role, joinedAt]
      properties:
        userId:
          type: string
        role:
          type: string
          enum: [owner, partner]
        joinedAt:
          type: string
          format: date-time
    Invite:
      type: object
      required: [id, code, status, expiresAt]
      properties:
        id:
          type: string
        code:
          type: string
        status:
          type: string
          enum: [issued, used, expired, revoked]
        expiresAt:
          type: string
          format: date-time